Update README.md
This project aims to create a visual dashboard to explore and browse geographic and geological boreholes’ data. This
is a joint project with the IGG institute of the CNR.
Specifically, we propose an application containing several visual widgets, each of which provides insights into a selection
of original data dimensions. The different visual widgets interact with each other, creating a series of connected displays
to browse data on multiple dimensions.
Interactivity was implemented by providing toolbars, selections and data filters. Moreover, we used dc.js library,
which is a javascript charting library with native crossfilter support, allowing highly efficient exploration on large multidimensional datasets. It leverages d3 to render charts in CSS-friendly SVG format. Charts rendered in this project are
data driven and reactive and therefore provide instant feedback to user interaction. The need to create such a dashboard
arises to ensure the interaction between the three data tables, object of our analysis.